refactor: extract getBreadcrumbItemSelector()
